Fix readme of unresolved merge conflict
Went with the readme content that occurs after the ======== merge
conflict indicator, which should be "the end of the original lines,
everything that follows (until the next indicator) is what's on
the branch that's being merged in."

For this readme file that is the content that started around line 73
to the end of the file.

Remove merge conflict indicators from file.